+++ internal/internal_task_handlers_test.go
@@ -125,9 +125,7 @@ func (t *TaskHandlersTestSuite) SetupSuite() {
 }
 
 func (t *TaskHandlersTestSuite) TearDownTest() {
-	if cache := *sharedWorkerCachePtr.workflowCache; cache != nil {
-		cache.Clear()
-	}
+	PurgeStickyWorkflowCache()
 }
 
 func TestTaskHandlersTestSuite(t *testing.T) {
@@ -517,7 +515,7 @@ func createTestEventTimerCanceled(eventID int64, id int) *historypb.HistoryEvent
 var testWorkflowTaskTaskqueue = "tq1"
 
 func (t *TaskHandlersTestSuite) getTestWorkerExecutionParams() workerExecutionParameters {
-	cache := NewWorkerCache()
+	cache := newTestWorkerCache(t.T())
 	return workerExecutionParameters{
 		TaskQueue:        testWorkflowTaskTaskqueue,
 		Namespace:        testNamespace,
@@ -1017,7 +1015,9 @@ func (t *TaskHandlersTestSuite) testSideEffectDeferHelper(cacheSize int) {
 	}
 
 	params := t.getTestWorkerExecutionParams()
-	params.cache = newWorkerCache(myWorkerCachePtr, &myWorkerCacheLock, cacheSize)
+	var cacheLease *workerCacheLease
+	params.cache, cacheLease = newWorkerCache(myWorkerCachePtr, &myWorkerCacheLock, cacheSize)
+	defer cacheLease.release()
 
 	taskHandler := newWorkflowTaskHandler(params, nil, t.registry)
 	task := createWorkflowTask(testEvents, 0, workflowName)
@@ -2690,7 +2690,8 @@ func TestResetIfDestroyedTaskPrep(t *testing.T) {
 			metricsHandler: metrics.NopHandler,
 			logger:         ilog.NewNopLogger(),
 			cache: &WorkerCache{
-				sharedCache: &sharedWorkerCache{workflowCache: &cache},
+				workflowCache:        cache,
+				maxWorkflowCacheSize: 1,
 			},
 		},
 	}