+++ internal/internal_workers_test.go
@@ -89,7 +89,7 @@ func (s *WorkersTestSuite) TestWorkflowWorker() {
 		BackgroundContext:       ctx,
 		BackgroundContextCancel: cancel,
 	}
-	overrides := &workerOverrides{workflowTaskHandler: newSampleWorkflowTaskHandler()}
+	overrides := &workerOverrides{workflowTaskHandler: newSampleWorkflowTaskHandler(s.T())}
 	client := &WorkflowClient{workflowService: s.service}
 	workflowWorker := newWorkflowWorkerInternal(client, executionParameters, nil, overrides, newRegistry())
 	s.Nil(workflowWorker.worker.options.taskPollers)
@@ -190,7 +190,7 @@ func (s *WorkersTestSuite) TestWorkflowWorkerSlotSupplier() {
 			Tuner:                   tuner,
 			WorkerStopTimeout:       time.Second,
 		}
-		overrides := &workerOverrides{workflowTaskHandler: newSampleWorkflowTaskHandler()}
+		overrides := &workerOverrides{workflowTaskHandler: newSampleWorkflowTaskHandler(s.T())}
 		client := &WorkflowClient{workflowService: s.service}
 		workflowWorker := newWorkflowWorkerInternal(client, executionParameters, nil, overrides, newRegistry())
 		_ = workflowWorker.Start()
@@ -450,7 +450,7 @@ func (s *WorkersTestSuite) TestPollWorkflowTaskQueue_InternalServiceError() {
 		),
 		Logger: ilog.NewNopLogger(),
 	}
-	overrides := &workerOverrides{workflowTaskHandler: newSampleWorkflowTaskHandler()}
+	overrides := &workerOverrides{workflowTaskHandler: newSampleWorkflowTaskHandler(s.T())}
 	client := &WorkflowClient{workflowService: s.service}
 	workflowWorker := newWorkflowWorkerInternal(client, executionParameters, nil, overrides, newRegistry())
 	_ = workflowWorker.Start()